+++ packages/acceptance-tests/pkg-tests-specs/sources/commands/workspace.test.ts
@@ -1,10 +1,17 @@
-const {
-  fs: {writeJson},
-} = require(`pkg-tests-core`);
+import {Filename, PortablePath, npath, ppath, xfs} from '@yarnpkg/fslib';
 
+async function setupWorkspaces(path: PortablePath) {
+  const docsFolder = `${path}/docs` as PortablePath;
 
-async function setupWorkspaces(path) {
-  await writeJson(`${path}/docs/package.json`, {
+  const componentAFolder = `${path}/components/component-a` as PortablePath;
+  const componentBFolder = `${path}/components/component-b` as PortablePath;
+
+  await xfs.mkdirPromise(docsFolder);
+
+  await xfs.mkdirPromise(componentAFolder, {recursive: true});
+  await xfs.mkdirPromise(componentBFolder, {recursive: true});
+
+  await xfs.writeJsonPromise(ppath.join(docsFolder, Filename.manifest), {
     name: `docs`,
     version: `1.0.0`,
     scripts: {
@@ -15,18 +22,19 @@ async function setupWorkspaces(path) {
     },
   });
 
-  await writeJson(`${path}/components/component-a/package.json`, {
+  await xfs.writeJsonPromise(ppath.join(componentAFolder, Filename.manifest), {
     name: `component-a`,
     version: `1.0.0`,
     scripts: {
       print: `echo Test Component A`,
+      printInitCwd: `echo $INIT_CWD`,
     },
     dependencies: {
       [`component-b`]: `workspace:*`,
     },
   });
 
-  await writeJson(`${path}/components/component-b/package.json`, {
+  await xfs.writeJsonPromise(ppath.join(componentBFolder, Filename.manifest), {
     name: `component-b`,
     version: `1.0.0`,
     scripts: {
@@ -64,6 +72,34 @@ describe(`Commands`, () => {
     );
 
     test(
+      `should preserve INIT_CWD to the original value`,
+      makeTemporaryEnv(
+        {
+          private: true,
+          workspaces: [`docs`, `components/*`],
+        },
+        async ({path, run}) => {
+          await setupWorkspaces(path);
+
+          let code;
+          let stdout;
+          let stderr;
+
+          try {
+            await run(`install`);
+            ({code, stdout, stderr} = await run(`workspace`, `component-a`, `run`, `printInitCwd`, {cwd: ppath.join(path, `docs` as PortablePath)}));
+          } catch (error) {
+            ({code, stdout, stderr} = error);
+          }
+
+          stdout = ppath.relative(path, npath.toPortablePath(stdout.trim()));
+
+          await expect({code, stdout, stderr}).toMatchSnapshot();
+        },
+      ),
+    );
+
+    test(
       `when the given workspace doesnt exist lists all possible workspaces`,
       makeTemporaryEnv(
         {
@@ -79,7 +115,7 @@ describe(`Commands`, () => {
 
           try {
             await run(`install`);
-            ({code, stdout, stderr} = await run(`workspace`, `compoment-a`, `print`));
+            ({code, stdout, stderr} = await run(`workspace`, `component-a`, `print`));
           } catch (error) {
             ({code, stdout, stderr} = error);
           }